Edit interface glow function, add exception handling, add new test for glow

+++ photoeffects/include/photoeffects.hpp
@@ -12,7 +12,7 @@ int fadeColor(cv::InputArray src, cv::OutputArray dst);
 
 cv::Mat tint(cv::Mat& image, int hue, float sat);
 
-int glow(cv::InputArray src, unsigned int radius, float intensity, cv::OutputArray dst);
+int glow(cv::InputArray src, float radius, float intensity, cv::OutputArray dst);
 
 int boost_color(cv::Mat& image);
